backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-dateutil==2.9.0
httpx==0.27.0
